+++ tests/test_simulations.py
@@ -506,6 +506,20 @@ class TestSimulation():
             simulation.data.synthetic[:, 7, 0].data
         )
 
+    def test_tqdm(self):
+        inp = {'survey': self.survey, 'model': self.model}
+
+        sim = simulations.Simulation(tqdm_opts=False, **inp)
+        assert sim._tqdm_opts['disable']
+
+        sim = simulations.Simulation(tqdm_opts=True, **inp)
+        assert not sim._tqdm_opts['disable']
+        assert sim._tqdm_opts['bar_format']
+
+        tqdm_opts = {'bar_format': '{bar}'}
+        sim = simulations.Simulation(tqdm_opts={'bar_format': '{bar}'}, **inp)
+        assert sim._tqdm_opts == tqdm_opts
+
 
 @pytest.mark.skipif(xarray is None, reason="xarray not installed.")
 def test_misfit():